net: shell: Add resume command
Move "resume" command to a separate file.
